apps/generate-api/requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
azure-functions>=1.20.0
pydantic>=2.8.0
